Added CPU/GPU features comparison to test
Originally, it wasn't possible to compare features extracted with a
CPU-based feature-extractor algorithm with features extracted with a
GPU-based feature-extractor algorithm. The reason for this was
differences in floating-point representations between CPU and GPU.

With the removal of the OpenCV upsampler, NVIDIA Optical Flow results
are returned as S10.5 floating-point values, which are much easier to
work with without incurring any odd floating-point representational
issues.

So this commit adds in a CPU-based feature-extractor algorithm to
confirm that the GPU-based feature-extractor algorithm is working
correctly.