Remove upsampling of Optical-Flow vectors (#5)
* Prototype of Optical-Flow upsampler removal

* Fixed issues with odd aspect ratios

* Fix incorrect issue submission URL

* Use in-tree feature-extractor kernel source file

* Added CPU/GPU features comparison to test

Originally, it wasn't possible to compare features extracted with a
CPU-based feature-extractor algorithm with features extracted with a
GPU-based feature-extractor algorithm. The reason for this was
differences in floating-point representations between CPU and GPU.

With the removal of the OpenCV upsampler, NVIDIA Optical Flow results
are returned as S10.5 floating-point values, which are much easier to
work with without incurring any odd floating-point representational
issues.

So this commit adds in a CPU-based feature-extractor algorithm to
confirm that the GPU-based feature-extractor algorithm is working
correctly.

* Quick Optical-Flow vector grid-size bug fix

* Fix test compiler warnings

* CI test fixes

Noticing some oddities between the GPU/CPU feature calculations between
the CI host and the development host machines. As a result, the test
conditions have beem loosened somewhat to accomodate for this.